requirements.txt
langchain-gigachat==0.x.y
langgraph==0.x.y
python-dotenv==0.x.y